refactor: create `trin-storage` workspace crate
