+++ docker/Dockerfile.bridge
@@ -23,6 +23,7 @@ COPY ./portal-bridge ./portal-bridge
 COPY ./trin-history ./trin-history
 COPY ./trin-metrics ./trin-metrics
 COPY ./trin-state ./trin-state
+COPY ./trin-storage ./trin-storage
 COPY ./trin-utils ./trin-utils
 COPY ./trin-validation ./trin-validation
 COPY ./ethportal-peertest ./ethportal-peertest
